app/requirements.txt
boto3==1.28.59
fastapi==0.103.2
fastapi-exceptionshandler==0.0.3
fastapi-versioning==0.10.0
opensearch-logger==1.2.3
pydantic==2.4.2
pydantic-settings==2.0.3
pyhumps==3.8.0
requests_aws4auth==1.2.3
SQLAlchemy==2.0.21
starlette==0.27.0
starlette-context==0.3.6